Fix cast_unittest warnings on win_x64_rel
Fix warnings with size_t casted to int.

TBR=justinlin@chromium.org

Review URL: https://codereview.chromium.org/26034003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@227051 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/cast/framer/cast_message_builder_unittest.cc
@@ -46,7 +46,8 @@ class NackFeedbackVerification : public RtpPayloadFeedback {
           std::make_pair(frame_it->first, kRtcpCastAllPacketsLost));
       } else {
       missing_packets_.insert(
-          std::make_pair(frame_it->first, frame_it->second.size()));
+          std::make_pair(frame_it->first,
+                         static_cast<int>(frame_it->second.size())));
       }
       ++frame_it;
     }
